# routes/swagger.yaml
openapi: 3.0.0
info:
  title: PocketMoney API
  version: 1.0.0
  description: API for PocketMoney application
servers:
  - url: 'http://localhost:3001'
    description: 'localhost'
  - url: 'http://34.170.36.196/api/'
    description: 'Google cloud'
paths:
...
  /budget/{id}:
    get:
      summary: Get budget detail
      description:
      tags:
        - budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: Budget ID
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
  /budget/new:
    post:
      summary: Add new budget
      tags:
        - budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  description: Name of the budget
                addMandatoryExpenses:
                  type: boolean
                  description: If true Mandatory expenses defined in configuration are added to the budget.
      responses:
        200:
          description: Success
  /budget/update/{id}:
    put:
      summary: Update existing budget
      description:
      tags:
        - budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget to be updated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  description: Name of the budget to be updated
      responses:
        200:
          description: Success
  /budget/delete/{id}:
    delete:
      summary: Delete budget
      description:
      tags:
        - budget
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            minimum: 1
          description: ID of the budget to be deleted
      responses:
        200:
          description: Success
...
components:
  schemas:
    Budget:
      type: object
      properties:
        id:
          type: object
          description: ID of the budget
        name:
          type: string
          description:
        balances:
          type: array
          description: List of balances in the budget
          items:
            $ref: '#components/schemas/BudgetBalance'
        expenses:
          type: array
          description: List of expected expenses in the budget
          items:
            $ref: '#components/schemas/BudgetExpense'
        balancesSums:
          $ref: '#components/schemas/BudgetBalancesSums'
...